+++ client/src/pages/EventPage/EventPageLayout.tsx
